refactor(core): update production type imports from coreToolScheduler
Updates imports in chatRecordingService.ts, geminiChat.ts, types.ts, and index.ts to use scheduler/types.ts.
