+++ packages/core/src/telemetry/loggers.test.circular.ts
@@ -12,11 +12,11 @@ import { describe, it, expect } from 'vitest';
 import { logToolCall } from './loggers.js';
 import { ToolCallEvent } from './types.js';
 import type { Config } from '../config/config.js';
-import type { CompletedToolCall } from '../core/coreToolScheduler.js';
 import {
   CoreToolCallStatus,
   type ToolCallRequestInfo,
   type ToolCallResponseInfo,
+  type CompletedToolCall,
 } from '../scheduler/types.js';
 import { MockTool } from '../test-utils/mock-tool.js';
